requirements.txt
discord.py[voice]==1.5.1
colour==0.1.5
